Add files via upload
This is the capstone project which was developed with the SQL , MS Excel , Glow root , Web stress , Selenium , JMeter , power point , word ,agile process , and JIRA. Here there are 4 main folders project management containing the Minutes of meetings , IAD log book , Meeting Agenda . The second folder is Reporting which contains the Testing report cycle for two phases . The third one is Testing which contains test scripts , test cases , traceability matrix and use cases developed in excel. The last one is the SQL file for the project.